Move old review models to legacy models file
