+++ api/approvals_api_test.py
@@ -23,7 +23,8 @@ from google.cloud import ndb  # type: ignore
 from api import approvals_api
 from internals import core_enums
 from internals import core_models
-from internals.review_models import Approval, ApprovalConfig, Gate, Vote
+from internals.legacy_models import Approval, ApprovalConfig
+from internals.review_models import Gate, Vote
 
 test_app = flask.Flask(__name__)
