+++ api/dev_api.py
@@ -18,8 +18,8 @@ from google.cloud import ndb  # type: ignore
 
 from framework.basehandlers import APIHandler
 from internals.core_models import FeatureEntry, MilestoneSet, Stage
-from internals.legacy_models import Feature
-from internals.review_models import Activity, Approval, Comment, Gate, Vote
+from internals.legacy_models import  Approval, Comment, Feature
+from internals.review_models import Activity, Gate, Vote
 from internals.core_enums import *
 import settings
