+++ internals/review_models.py
@@ -23,154 +23,6 @@ from typing import Optional
 from google.cloud import ndb  # type: ignore
 
 
-class Approval(ndb.Model):
-  """Describes the current state of one approval on a feature."""
-
-  # Not used: PREPARING = 0
-  NA = 1
-  REVIEW_REQUESTED = 2
-  REVIEW_STARTED = 3
-  NEEDS_WORK = 4
-  APPROVED = 5
-  DENIED = 6
-  NO_RESPONSE = 7
-  INTERNAL_REVIEW = 8
-  APPROVAL_VALUES = {
-      # Not used: PREPARING: 'preparing',
-      NA: 'na',
-      REVIEW_REQUESTED: 'review_requested',
-      REVIEW_STARTED: 'review_started',
-      NEEDS_WORK: 'needs_work',
-      APPROVED: 'approved',
-      DENIED: 'denied',
-      NO_RESPONSE: 'no_response',
-      INTERNAL_REVIEW: 'internal_review',
-  }
-
-  FINAL_STATES = [NA, APPROVED, DENIED]
-
-  feature_id = ndb.IntegerProperty(required=True)
-  field_id = ndb.IntegerProperty(required=True)
-  state = ndb.IntegerProperty(required=True)
-  set_on = ndb.DateTimeProperty(required=True)
-  set_by = ndb.StringProperty(required=True)
-
-  @classmethod
-  def get_approvals(
-      cls, feature_id=None, field_id=None, states=None, set_by=None,
-      limit=None) -> list[Approval]:
-    """Return the requested approvals."""
-    query = Approval.query().order(Approval.set_on)
-    if feature_id is not None:
-      query = query.filter(Approval.feature_id == feature_id)
-    if field_id is not None:
-      query = query.filter(Approval.field_id == field_id)
-    if states is not None:
-      query = query.filter(Approval.state.IN(states))
-    if set_by is not None:
-      query = query.filter(Approval.set_by == set_by)
-    # Query with STRONG consistency because ndb defaults to
-    # EVENTUAL consistency and we run this query immediately after
-    # saving the user's change that we want included in the query.
-    approvals = query.fetch(limit, read_consistency=ndb.STRONG)
-    return approvals
-
-  @classmethod
-  def is_valid_state(cls, new_state):
-    """Return true if new_state is valid."""
-    return new_state in cls.APPROVAL_VALUES
-
-  @classmethod
-  def set_approval(cls, feature_id, field_id, new_state, set_by_email):
-    """Add or update an approval value."""
-    if not cls.is_valid_state(new_state):
-      raise ValueError('Invalid approval state')
-
-    now = datetime.datetime.now()
-    existing_list = cls.get_approvals(
-        feature_id=feature_id, field_id=field_id, set_by=set_by_email)
-    if existing_list:
-      existing = existing_list[0]
-      existing.set_on = now
-      existing.state = new_state
-      existing.put()
-      logging.info('existing approval is %r', existing.key.integer_id())
-      return
-
-    new_appr = Approval(
-        feature_id=feature_id, field_id=field_id, state=new_state,
-        set_on=now, set_by=set_by_email)
-    new_appr.put()
-    logging.info('new_appr is %r', new_appr.key.integer_id())
-
-  @classmethod
-  def clear_request(cls, feature_id, field_id):
-    """After the review requirement has been satisfied, remove the request."""
-    review_requests = cls.get_approvals(
-        feature_id=feature_id, field_id=field_id, states=[cls.REVIEW_REQUESTED])
-    for rr in review_requests:
-      rr.key.delete()
-
-    # Note: We keep REVIEW_REQUEST Vote entities.
-
-
-class ApprovalConfig(ndb.Model):
-  """Allows customization of an approval field for one feature."""
-
-  feature_id = ndb.IntegerProperty(required=True)
-  field_id = ndb.IntegerProperty(required=True)
-  owners = ndb.StringProperty(repeated=True)
-  next_action = ndb.DateProperty()
-  additional_review = ndb.BooleanProperty(default=False)
-
-  @classmethod
-  def get_configs(cls, feature_id):
-    """Return approval configs for all approval fields."""
-    query = ApprovalConfig.query(ApprovalConfig.feature_id == feature_id)
-    configs = query.fetch(None)
-    return configs
-
-  @classmethod
-  def set_config(
-      cls, feature_id, field_id, owners, next_action, additional_review):
-    """Add or update an approval config object."""
-    config = ApprovalConfig(feature_id=feature_id, field_id=field_id)
-    for existing in cls.get_configs(feature_id):
-      if existing.field_id == field_id:
-        config = existing
-
-    config.owners = owners or []
-    config.next_action = next_action
-    config.additional_review = additional_review
-    config.put()
-
-
-class Comment(ndb.Model):
-  """A review comment on a feature."""
-  feature_id = ndb.IntegerProperty(required=True)
-  field_id = ndb.IntegerProperty()  # The approval field_id, or general comment.
-  created = ndb.DateTimeProperty(auto_now_add=True)
-  author = ndb.StringProperty()
-  content = ndb.StringProperty()
-  deleted_by = ndb.StringProperty()
-  migrated = ndb.BooleanProperty()
-  # If the user set an approval value, we capture that here so that we can
-  # display a change log.  This could be generalized to a list of separate
-  # Amendment entities, but that complexity is not needed yet.
-  old_approval_state = ndb.IntegerProperty()
-  new_approval_state = ndb.IntegerProperty()
-
-  @classmethod
-  def get_comments(cls, feature_id, field_id=None):
-    """Return review comments for an approval."""
-    query = Comment.query().order(Comment.created)
-    query = query.filter(Comment.feature_id == feature_id)
-    if field_id:
-      query = query.filter(Comment.field_id == field_id)
-    comments = query.fetch(None)
-    return comments
-
-
 class OwnersFile(ndb.Model):
   """Describes the properties to store raw API_OWNERS content."""
   url = ndb.StringProperty(required=True)
@@ -203,7 +55,7 @@ class OwnersFile(ndb.Model):
     return owners_file.raw_content
 
 
-class Vote(ndb.Model):  # copy from Approval
+class Vote(ndb.Model):
   """One approver's vote on what the state of a gate should be."""
 
   # Not used: PREPARING = 0
@@ -267,7 +119,7 @@ class Vote(ndb.Model):  # copy from Approval
   # Note: set_vote() moved to approval_defs.py
 
 
-class Gate(ndb.Model):  # copy from ApprovalConfig
+class Gate(ndb.Model):
   """Gates regulate the completion of a stage."""
 
   PREPARING = 0
