+++ internals/search_queries_test.py
@@ -18,8 +18,8 @@ import datetime
 from unittest import mock
 
 from internals import core_enums
-from internals import core_models
-from internals import review_models
+from internals.core_models import FeatureEntry, MilestoneSet, Stage
+from internals.review_models import Gate, Vote
 from internals import search
 from internals import search_queries
 
@@ -27,7 +27,7 @@ from internals import search_queries
 class SearchFeaturesTest(testing_config.CustomTestCase):
 
   def setUp(self):
-    self.feature_1 = core_models.FeatureEntry(
+    self.feature_1 = FeatureEntry(
         name='feature a', summary='sum',
         category=1, impl_status_chrome=3)
     self.feature_1.owner_emails = ['owner@example.com']
@@ -36,68 +36,68 @@ class SearchFeaturesTest(testing_config.CustomTestCase):
     self.feature_1.put()
     self.feature_1_id = self.feature_1.key.integer_id()
 
-    self.stage_1_ship = core_models.Stage(
+    self.stage_1_ship = Stage(
         feature_id=self.feature_1_id,
         stage_type=core_enums.STAGE_BLINK_SHIPPING,
-        milestones=core_models.MilestoneSet(desktop_first=99))
+        milestones=MilestoneSet(desktop_first=99))
     self.stage_1_ship.put()
 
-    self.feature_2 = core_models.FeatureEntry(
+    self.feature_2 = FeatureEntry(
         name='feature b', summary='sum', owner_emails=['owner@example.com'],
         category=1, impl_status_chrome=3)
     self.feature_2.put()
     self.feature_2_id = self.feature_2.key.integer_id()
 
-    self.feature_3 = core_models.FeatureEntry(
+    self.feature_3 = FeatureEntry(
         name='feature c', summary='sum', owner_emails=['random@example.com'],
         category=1, impl_status_chrome=4)
     self.feature_3.put()
     self.feature_3_id = self.feature_3.key.integer_id()
 
-    self.gate_1 = review_models.Gate(
+    self.gate_1 = Gate(
         feature_id=self.feature_1_id, stage_id=1,
-        gate_type=core_models.GATE_API_PROTOTYPE,
-        state=review_models.Vote.APPROVED,
+        gate_type=core_enums.GATE_API_PROTOTYPE,
+        state=Vote.APPROVED,
         requested_on=datetime.datetime(2022, 7, 1))
     self.gate_1.put()
     self.gate_1_id = self.gate_1.key.integer_id()
 
-    self.vote_1_1 = review_models.Vote(
-        feature_id=self.feature_1_id, gate_type=core_models.GATE_API_PROTOTYPE,
+    self.vote_1_1 = Vote(
+        feature_id=self.feature_1_id, gate_type=core_enums.GATE_API_PROTOTYPE,
         gate_id=self.gate_1_id,
-        state=review_models.Vote.REVIEW_REQUESTED,
+        state=Vote.REVIEW_REQUESTED,
         set_on=datetime.datetime(2022, 7, 1),
         set_by='feature_owner@example.com')
     self.vote_1_1.put()
 
-    self.vote_1_2 = review_models.Vote(
-        feature_id=self.feature_1_id, gate_type=core_models.GATE_API_PROTOTYPE,
+    self.vote_1_2 = Vote(
+        feature_id=self.feature_1_id, gate_type=core_enums.GATE_API_PROTOTYPE,
         gate_id=self.gate_1_id,
-        state=review_models.Vote.APPROVED,
+        state=Vote.APPROVED,
         set_on=datetime.datetime(2022, 7, 2),
         set_by='reviewer@example.com')
     self.vote_1_2.put()
 
-    self.gate_2 = review_models.Gate(
+    self.gate_2 = Gate(
         feature_id=self.feature_2_id, stage_id=1,
         gate_type=core_enums.GATE_API_SHIP,
-        state=review_models.Vote.REVIEW_REQUESTED,
+        state=Vote.REVIEW_REQUESTED,
         requested_on=datetime.datetime(2022, 8, 1))
     self.gate_2.put()
     self.gate_2_id = self.gate_2.key.integer_id()
 
-    self.vote_2_1 = review_models.Vote(
+    self.vote_2_1 = Vote(
         feature_id=self.feature_2_id, gate_type=core_enums.GATE_API_SHIP,
         gate_id=self.gate_2_id,
-        state=review_models.Vote.REVIEW_REQUESTED,
+        state=Vote.REVIEW_REQUESTED,
         set_on=datetime.datetime(2022, 8, 1),
         set_by='feature_owner@example.com')
     self.vote_2_1.put()
 
-    self.vote_2_2 = review_models.Vote(
+    self.vote_2_2 = Vote(
         feature_id=self.feature_2_id, gate_type=core_enums.GATE_API_SHIP,
         gate_id=self.gate_2_id,
-        state=review_models.Vote.APPROVED,
+        state=Vote.APPROVED,
         set_on=datetime.datetime(2022, 8, 2),
         set_by='reviewer@example.com')
     self.vote_2_2.put()
@@ -107,9 +107,9 @@ class SearchFeaturesTest(testing_config.CustomTestCase):
     self.feature_2.key.delete()
     self.feature_3.key.delete()
     self.stage_1_ship.key.delete()
-    for gate in review_models.Gate.query():
+    for gate in Gate.query():
       gate.key.delete()
-    for vote in review_models.Vote.query():
+    for vote in Vote.query():
       vote.key.delete()
 
   def test_single_field_query_async__normal(self):
